Run ReactElementJSX-test against bundles
